Computer-Vision/requirements.txt
numpy==1.19.5
pandas==1.2.4
Pillow==8.4.0
scikit_learn==1.0.1
tensorflow==2.3.0
